git-svn-id: http://ovgl.googlecode.com/svn/trunk@74 f7c3e6a6-3a5d-87ca-2239-9ed928e7452e

+++ include/OvglAudio.h
@@ -42,7 +42,7 @@ namespace Ovgl
 			int									frequency;
 			uint32_t							stereo;
 			uint32_t							mono;
-            std::vector< int16_t >				data;
+            std::vector< int8_t >				data;
 			AudioInstance* CreateAudioInstance( Emitter* emitter, bool loop );
 			void Release();
 		};
@@ -64,7 +64,7 @@ namespace Ovgl
 		{
 		public:
 			AudioInstance*						instance;
-			uint32_t						source;
+            uint32_t                            source;
 			void Release();
 		};
 	}
